// src/app/editArticle/types/editArticleState.inerface.ts
import { ArticleInterface } from "src/app/shared/types/article.interface";
import { BackendErrorsInterface } from "src/app/shared/types/backendErrors.interface";

export interface EditArticleStateInterface {
   isLoading: boolean;
   isSubmitting: boolean;
   validationErrors: BackendErrorsInterface | null;
   article: ArticleInterface | null;
}